backend/pyproject.toml
[tool.poetry]
name = "bitrender"
version = "0.1.0"
description = ""
authors = ["Robin Schyboll <Robin.Schyboll@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.74.1"
uvicorn = "^0.17.5"
asyncpg = "^0.25.0"
aerich = "^0.6.2"
tortoise-orm = { git = "https://github.com/Hoodrobinrs/tortoise-orm.git", branch = "generic-models" }
bcrypt = "^3.2.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}

[tool.poetry.dev-dependencies]
pytest = "^7.0.1"
mypy = "^0.931"
flake8 = "^4.0.1"
black = "^22.1.0"
pylint = "^2.12.2"
isort = "^5.10.1"
Sphinx = "^4.4.0"
pytest-mock = "^3.7.0"
pytest-asyncio = "^0.18.1"
pytest-cov = "^3.0.0"
